+++ crates/burn-central-runtime/src/lib.rs
@@ -1,3 +1,8 @@
+mod error;
+mod routine;
+mod param;
+mod types;
+mod output;
 mod executor;
 mod type_name;
 
@@ -6,7 +11,10 @@ mod backend;
 #[cfg(feature = "cli")]
 pub mod cli;
 
-pub use executor::*;
+pub use executor::{ExecutorBuilder, Executor, ExecutionContext};
+pub use error::RuntimeError;
+pub use routine::{IntoRoutine, Routine};
+pub use types::*;
 
 pub fn setup_logging() {
     env_logger::builder()
